Some code refactor
